+++ makefile
@@ -1,19 +1,19 @@
 a.out:  My_Point2d.o My_Group2D.o My_Cluster.o My_Poisk.o My_Field.o My_Controller.o My_Interface.o main.o
 	g++ My_Point2d.o My_Group2D.o My_Cluster.o My_Poisk.o My_Field.o My_Controller.o My_Interface.o main.o
 
-My_Field.o: My_Field.cpp My_Field.h
+My_Field.o: My_Field.cpp My_Field.h My_Classes.h
 	g++ -c My_Field.cpp
 
-My_Point2d.o: My_Point2d.cpp My_Point2d.h
+My_Point2d.o: My_Point2d.cpp My_Classes.h
 	g++ -c My_Point2d.cpp
 
-My_Group2D.o: My_Group2D.cpp My_Group2D.h
+My_Group2D.o: My_Group2D.cpp My_Classes.h
 	g++ -c My_Group2D.cpp
 
-My_Cluster.o: My_Cluster.cpp My_Cluster.h
+My_Cluster.o: My_Cluster.cpp My_Classes.h
 	g++ -c My_Cluster.cpp
 
-My_Poisk.o: My_Poisk.cpp My_Poisk.h
+My_Poisk.o: My_Poisk.cpp My_Classes.h
 	g++ -c My_Poisk.cpp
 
 My_Controller.o: My_Controller.cpp My_Controller.h
